Added getElement to postParsing from its own file.
